Add Rock Paper Scissors game and IDE/config files to .gitignore
This commit includes a new Rock Paper Scissors game with a player class and runner script. The game is constructed with a player and a random computer choice, and determines the winner based on these choices.

Also, several configuration files from the IntelliJ IDEA (.idea folder, misc.xml, vcs.xml, modules.xml, and the .iml project file) have been added to the repository, as well as MouseButtonTester.java from a different project. The purpose of these changes is to correctly setup a development environment. However, IDE specific files that don't need to be shared with other developers were added to .gitignore to avoid cluttering the repository.