Add MinMax class and Runner along with .xml configurations
Added new MinMax class for calculating Y and G values, and a runner class to output these calculations. XML configuration files were updated and added to support this new functionality and organize the project structure. Various development and environment configuration files are also added to the repository to improve portability and collaborator convenience on this project.