Add Window1 and subsequent Window2 class with interactive color switch buttons
This commit introduces a new class, Window2, in the window2 package. This class extends the functionalities of the previously implemented window1 class by adding interactive buttons, which enables the user to set the background color of the window to red or green. Improve user interactivity and provide a demonstration of the practical use of action listeners in a GUI.